+++ lib/openapi_blocks/controller.rb
@@ -2,8 +2,10 @@
 
 module OpenapiBlocks
   class Controller # rubocop:disable Style/Documentation
+    include Concerns::Documentable
+
     class << self
-      attr_reader :_resource, :_operations, :_tags, :_controller_class
+      attr_reader :_resource, :_controller_class
 
       def resource(klass)
         @_resource = klass
@@ -13,32 +15,10 @@ module OpenapiBlocks
         @_controller_class = klass
       end
 
-      def model
-        @_resource&.model
-      end
-
-      def _associations
-        @_resource&._associations
-      end
-
-      def _virtual_attributes
-        @_resource&._virtual_attributes
-      end
-
-      def _ignored
-        @_resource&._ignored
-      end
-
-      def operation(action, &block)
-        @_operations ||= {}
-        builder = OperationBuilder.new
-        builder.instance_eval(&block) if block
-        @_operations[action] = builder
-      end
-
-      def tags(*values)
-        values.any? ? @_tags = values : @_tags
-      end
+      def model               = @_resource&.model
+      def _associations       = @_resource&._associations
+      def _virtual_attributes = @_resource&._virtual_attributes
+      def _ignored            = @_resource&._ignored
     end
   end
 end